Move menu building related code into BVC extension
