+++ Client/Extensions/AppearanceExtensions.swift
@@ -43,7 +43,6 @@ extension Theme {
         UITableViewCell.appearance().backgroundColor = tableSecondaryColor
         
         UIImageView.appearance(whenContainedInInstancesOf: [SettingsViewController.self]).tintColor = colors.tints.home
-//        UIImageView.appearance(whenContainedInInstancesOf: [NewMenuHostingController.self]).tintColor = colors.tints.home
         UIImageView.appearance(whenContainedInInstancesOf: [BraveRewardsSettingsViewController.self]).tintColor = colors.tints.home
 
         UIView.appearance(whenContainedInInstancesOf: [UITableViewHeaderFooterView.self]).appearanceBackgroundColor = tablePrimaryColor
